+++ internal/provider/provider.go
@@ -79,6 +79,7 @@ type providerData struct {
 	EnablePreflight              types.Bool   `tfsdk:"enable_preflight"`
 	DisableDefaultOutput         types.Bool   `tfsdk:"disable_default_output"`
 	CustomHeaders                types.Map    `tfsdk:"custom_headers"`
+	MaximumBusyRetryAttempts     types.Int32  `tfsdk:"maximum_busy_retry_attempts"`
 }
 
 func (model providerData) GetClientId() (*string, error) {
@@ -196,17 +197,17 @@ func (p Provider) Schema(ctx context.Context, request provider.SchemaRequest, re
 					Attributes: map[string]schema.Attribute{
 						"active_directory_authority_host": schema.StringAttribute{
 							Optional:            true,
-							MarkdownDescription: "The Azure Resource Manager endpoint to use. This can also be sourced from the `ARM_RESOURCE_MANAGER_ENDPOINT` Environment Variable. Defaults to `https://management.azure.com/` for public cloud.",
+							MarkdownDescription: "The Azure Active Directory login endpoint to use. This can also be sourced from the `ARM_ACTIVE_DIRECTORY_AUTHORITY_HOST` Environment Variable. Defaults to `https://login.microsoftonline.com/` for public cloud.",
 						},
 
 						"resource_manager_endpoint": schema.StringAttribute{
 							Optional:            true,
-							MarkdownDescription: "The resource ID to obtain AD tokens for. This can also be sourced from the `ARM_RESOURCE_MANAGER_AUDIENCE` Environment Variable. Defaults to `https://management.core.windows.net/` for public cloud.",
+							MarkdownDescription: "The Azure Resource Manager endpoint to use. This can also be sourced from the `ARM_RESOURCE_MANAGER_ENDPOINT` Environment Variable. Defaults to `https://management.azure.com/` for public cloud.",
 						},
 
 						"resource_manager_audience": schema.StringAttribute{
 							Optional:            true,
-							MarkdownDescription: "The Azure Active Directory login endpoint to use. This can also be sourced from the `ARM_ACTIVE_DIRECTORY_AUTHORITY_HOST` Environment Variable. Defaults to `https://login.microsoftonline.com/` for public cloud.",
+							MarkdownDescription: "The resource ID to obtain AD tokens for. This can also be sourced from the `ARM_RESOURCE_MANAGER_AUDIENCE` Environment Variable. Defaults to `https://management.core.windows.net/` for public cloud.",
 						},
 					},
 				},
@@ -263,12 +264,12 @@ func (p Provider) Schema(ctx context.Context, request provider.SchemaRequest, re
 			// OIDC specific fields
 			"oidc_request_token": schema.StringAttribute{
 				Optional:            true,
-				MarkdownDescription: "The bearer token for the request to the OIDC provider. This can also be sourced from the `ARM_OIDC_REQUEST_TOKEN` or `ACTIONS_ID_TOKEN_REQUEST_TOKEN` Environment Variables.",
+				MarkdownDescription: "The bearer token for the request to the OIDC provider. This can also be sourced from the `ARM_OIDC_REQUEST_TOKEN`, `ACTIONS_ID_TOKEN_REQUEST_TOKEN`, or `SYSTEM_ACCESSTOKEN` Environment Variables.",
 			},
 
 			"oidc_request_url": schema.StringAttribute{
 				Optional:            true,
-				MarkdownDescription: "The URL for the OIDC provider from which to request an ID token. This can also be sourced from the `ARM_OIDC_REQUEST_URL` or `ACTIONS_ID_TOKEN_REQUEST_URL` Environment Variables.",
+				MarkdownDescription: "The URL for the OIDC provider from which to request an ID token. This can also be sourced from the `ARM_OIDC_REQUEST_URL`, `ACTIONS_ID_TOKEN_REQUEST_URL`, or `SYSTEM_OIDCREQUESTURI` Environment Variables.",
 			},
 
 			"oidc_token": schema.StringAttribute{
@@ -283,7 +284,7 @@ func (p Provider) Schema(ctx context.Context, request provider.SchemaRequest, re
 
 			"oidc_azure_service_connection_id": schema.StringAttribute{
 				Optional:            true,
-				MarkdownDescription: "The Azure Pipelines Service Connection ID to use for authentication. This can also be sourced from the `ARM_ADO_PIPELINE_SERVICE_CONNECTION_ID` or `ARM_OIDC_AZURE_SERVICE_CONNECTION_ID` Environment Variables.",
+				MarkdownDescription: "The Azure Pipelines Service Connection ID to use for authentication. This can also be sourced from the `ARM_ADO_PIPELINE_SERVICE_CONNECTION_ID`, `ARM_OIDC_AZURE_SERVICE_CONNECTION_ID`, or `AZURESUBSCRIPTION_SERVICE_CONNECTION_ID` Environment Variables.",
 			},
 
 			"use_oidc": schema.BoolAttribute{
@@ -361,12 +362,17 @@ func (p Provider) Schema(ctx context.Context, request provider.SchemaRequest, re
 
 			"enable_preflight": schema.BoolAttribute{
 				Optional:    true,
-				Description: "Enable Preflight Validation. The default is false. When set to true, the provider will use Preflight to do static validation before really deploying a new resource. When set to false, the provider will disable this validation.",
+				Description: "Enable Preflight Validation. The default is false. When set to true, the provider will use Preflight to do static validation before really deploying a new resource. When set to false, the provider will disable this validation. This can also be sourced from the `ARM_ENABLE_PREFLIGHT` Environment Variable.",
 			},
 
 			"disable_default_output": schema.BoolAttribute{
 				Optional:    true,
-				Description: "Disable default output. The default is false. When set to false, the provider will output the read-only properties if `response_export_values` is not specified in the resource block. When set to true, the provider will disable this output.",
+				Description: "Disable default output. The default is false. When set to false, the provider will output the read-only properties if `response_export_values` is not specified in the resource block. When set to true, the provider will disable this output. This can also be sourced from the `ARM_DISABLE_DEFAULT_OUTPUT` Environment Variable.",
+			},
+
+			"maximum_busy_retry_attempts": schema.Int32Attribute{
+				Optional:            true,
+				MarkdownDescription: "The maximum number of retries to attempt if the Azure API returns an HTTP 408, 429, 500, 502, 503, or 504 response. The default is `3`. The resource-specific retry configuration may additionally be used to retry on other errors and conditions.",
 			},
 
 			"custom_headers": schema.MapAttribute{
@@ -503,6 +509,8 @@ func (p Provider) Configure(ctx context.Context, request provider.ConfigureReque
 			model.OIDCRequestToken = types.StringValue(v)
 		} else if v := os.Getenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN"); v != "" {
 			model.OIDCRequestToken = types.StringValue(v)
+		} else if v := os.Getenv("SYSTEM_ACCESSTOKEN"); v != "" {
+			model.OIDCRequestToken = types.StringValue(v)
 		}
 	}
 
@@ -530,6 +538,8 @@ func (p Provider) Configure(ctx context.Context, request provider.ConfigureReque
 		if v := os.Getenv("ARM_ADO_PIPELINE_SERVICE_CONNECTION_ID"); v != "" {
 			model.OIDCAzureServiceConnectionID = types.StringValue(v)
 		} else if v := os.Getenv("ARM_OIDC_AZURE_SERVICE_CONNECTION_ID"); v != "" {
+			model.OIDCAzureServiceConnectionID = types.StringValue(v)
+		} else if v := os.Getenv("AZURESUBSCRIPTION_SERVICE_CONNECTION_ID"); v != "" {
 			model.OIDCAzureServiceConnectionID = types.StringValue(v)
 		}
 	}
@@ -587,10 +597,18 @@ func (p Provider) Configure(ctx context.Context, request provider.ConfigureReque
 	}
 
 	if model.EnablePreflight.IsNull() {
-		model.EnablePreflight = types.BoolValue(false)
+		if v := os.Getenv("ARM_ENABLE_PREFLIGHT"); v != "" {
+			model.EnablePreflight = types.BoolValue(v == "true")
+		} else {
+			model.EnablePreflight = types.BoolValue(false)
+		}
 	}
 	if model.DisableDefaultOutput.IsNull() {
-		model.DisableDefaultOutput = types.BoolValue(false)
+		if v := os.Getenv("ARM_DISABLE_DEFAULT_OUTPUT"); v != "" {
+			model.DisableDefaultOutput = types.BoolValue(v == "true")
+		} else {
+			model.DisableDefaultOutput = types.BoolValue(false)
+		}
 	}
 
 	var cloudConfig cloud.Configuration
@@ -660,10 +678,16 @@ func (p Provider) Configure(ctx context.Context, request provider.ConfigureReque
 		customHeaders[k] = []string{v.(basetypes.StringValue).ValueString()}
 	}
 
+	maxGoSdkRetryAttempts := int32(3)
+	if !model.MaximumBusyRetryAttempts.IsNull() {
+		maxGoSdkRetryAttempts = model.MaximumBusyRetryAttempts.ValueInt32()
+	}
+
 	copt := &clients.Option{
 		Cred:                 cred,
 		CloudCfg:             cloudConfig,
 		ApplicationUserAgent: buildUserAgent(request.TerraformVersion, model.PartnerID.ValueString(), model.DisableTerraformPartnerID.ValueBool()),
+		MaxGoSdkRetries:      maxGoSdkRetryAttempts,
 		Features: features.UserFeatures{
 			DefaultTags:          tags.ExpandTags(model.DefaultTags),
 			DefaultLocation:      location.Normalize(model.DefaultLocation.ValueString()),
@@ -708,6 +732,7 @@ func (p Provider) Functions(ctx context.Context) []func() function.Function {
 		func() function.Function { return &functions.ResourceGroupResourceIdFunction{} },
 		func() function.Function { return &functions.ManagementGroupResourceIdFunction{} },
 		func() function.Function { return &functions.ExtensionResourceIdFunction{} },
+		func() function.Function { return &functions.UniqueStringFunction{} },
 	}
 }