+++ internal/services/azapi_resource.go
@@ -30,7 +30,6 @@ import (
 	"github.com/Azure/terraform-provider-azapi/internal/skip"
 	"github.com/Azure/terraform-provider-azapi/internal/tf"
 	"github.com/Azure/terraform-provider-azapi/utils"
-	"github.com/cenkalti/backoff/v4"
 	"github.com/hashicorp/terraform-plugin-framework-timeouts/resource/timeouts"
 	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
@@ -52,6 +51,7 @@ const FlagMoveState = "move_state"
 
 type AzapiResourceModel struct {
 	Body                          types.Dynamic    `tfsdk:"body"`
+	SensitiveBody                 types.Dynamic    `tfsdk:"sensitive_body"`
 	ID                            types.String     `tfsdk:"id"`
 	Identity                      types.List       `tfsdk:"identity"`
 	IgnoreCasing                  types.Bool       `tfsdk:"ignore_casing"`
@@ -117,7 +117,7 @@ func (r *AzapiResource) Schema(ctx context.Context, _ resource.SchemaRequest, re
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.UseStateForUnknown(),
 				},
-				MarkdownDescription: docstrings.Type(),
+				MarkdownDescription: docstrings.ID(),
 			},
 
 			"name": schema.StringAttribute{
@@ -173,6 +173,12 @@ func (r *AzapiResource) Schema(ctx context.Context, _ resource.SchemaRequest, re
 				Validators: []validator.Dynamic{
 					myvalidator.DynamicIsNotStringValidator(),
 				},
+			},
+
+			"sensitive_body": schema.DynamicAttribute{
+				Optional:            true,
+				WriteOnly:           true,
+				MarkdownDescription: docstrings.SensitiveBody(),
 			},
 
 			"replace_triggers_external_values": schema.DynamicAttribute{
@@ -265,7 +271,7 @@ func (r *AzapiResource) Schema(ctx context.Context, _ resource.SchemaRequest, re
 				MarkdownDescription: "A mapping of tags which should be assigned to the Azure resource.",
 			},
 
-			"retry": retry.SingleNestedAttribute(ctx),
+			"retry": retry.RetrySchema(ctx),
 
 			"create_headers": schema.MapAttribute{
 				ElementType:         types.StringType,
@@ -382,6 +388,9 @@ func (r *AzapiResource) ValidateConfig(ctx context.Context, request resource.Val
 	if config == nil {
 		return
 	}
+	if config.Type.IsUnknown() {
+		return
+	}
 
 	resourceType := config.Type.ValueString()
 
@@ -394,19 +403,20 @@ func (r *AzapiResource) ValidateConfig(ctx context.Context, request resource.Val
 		}
 	}
 
-	if !dynamic.IsFullyKnown(config.Body) {
-		return
-	}
-
-	body := make(map[string]interface{})
-	if err := unmarshalBody(config.Body, &body); err != nil {
-		response.Diagnostics.AddError("Invalid body", fmt.Sprintf(`The argument "body" is invalid: %s`, err.Error()))
-		return
-	}
-
-	if diags := validateDuplicatedDefinitions(config, body); diags.HasError() {
+	if diags := validateDuplicatedDefinitions(config, config.Body, "body"); diags.HasError() {
 		response.Diagnostics.Append(diags...)
 		return
+	}
+	if diags := validateDuplicatedDefinitions(config, config.SensitiveBody, "sensitive_body"); diags.HasError() {
+		response.Diagnostics.Append(diags...)
+		return
+	}
+
+	if config.SchemaValidationEnabled.IsNull() || config.SchemaValidationEnabled.ValueBool() {
+		if err := schemaValidate(config); err != nil {
+			response.Diagnostics.AddError("Invalid configuration", err.Error())
+			return
+		}
 	}
 }
 
@@ -481,42 +491,27 @@ func (r *AzapiResource) ModifyPlan(ctx context.Context, request resource.ModifyP
 		}
 	}
 
-	if dynamic.IsFullyKnown(plan.Body) {
-		body := make(map[string]interface{})
-		if err := unmarshalBody(config.Body, &body); err != nil {
-			response.Diagnostics.AddError("Invalid body", fmt.Sprintf(`The argument "body" is invalid: %s`, err.Error()))
-			return
-		}
+	// Set output as unknown to trigger a plan diff, if ephemral body has changed
+	diff, diags := ephemeralBodyChangeInPlan(ctx, request.Private, config.SensitiveBody)
+	if response.Diagnostics = append(response.Diagnostics, diags...); response.Diagnostics.HasError() {
+		return
+	}
+	if diff {
+		tflog.Info(ctx, `"sensitive_body" has changed`)
+		plan.Output = types.DynamicUnknown()
+	}
 
-		plan.Tags = r.tagsWithDefaultTags(config.Tags, body, state, resourceDef)
+	if dynamic.IsFullyKnown(plan.Body) {
+		plan.Tags = r.tagsWithDefaultTags(config.Tags, state, config.Body, resourceDef)
 		if state == nil || !state.Tags.Equal(plan.Tags) {
 			plan.Output = basetypes.NewDynamicUnknown()
 		}
 
-		// location field has a field level plan modifier which suppresses the diff if the location is not actually changed
-		locationValue := plan.Location
-		// For the following cases, we need to use the location in config as the specified location
-		// case 1. To create a new resource, the location is not specified in config, then the planned location will be unknown
-		// case 2. To update a resource, the location is not specified in config, then the planned location will be the state location
-		if locationValue.IsUnknown() || config.Location.IsNull() {
-			locationValue = config.Location
-		}
 		// locationWithDefaultLocation will return the location in config if it's not null, otherwise it will return the default location if it supports location
-		plan.Location = r.locationWithDefaultLocation(locationValue, body, state, resourceDef)
+		plan.Location = r.locationWithDefaultLocation(config.Location, plan.Location, state, config.Body, resourceDef)
 		if state != nil && location.Normalize(state.Location.ValueString()) != location.Normalize(plan.Location.ValueString()) {
 			// if the location is changed, replace the resource
 			response.RequiresReplace.Append(path.Root("location"))
-		}
-		if plan.SchemaValidationEnabled.ValueBool() {
-			if response.Diagnostics.Append(expandBody(body, *plan)...); response.Diagnostics.HasError() {
-				return
-			}
-			body["name"] = plan.Name.ValueString()
-			err = schemaValidation(azureResourceType, apiVersion, resourceDef, body)
-			if err != nil {
-				response.Diagnostics.AddError("Invalid configuration", err.Error())
-				return
-			}
 		}
 
 		// Check if any paths in replace_triggers_refs have changed
@@ -561,7 +556,7 @@ func (r *AzapiResource) ModifyPlan(ctx context.Context, request resource.ModifyP
 		}
 	}
 
-	if r.ProviderData.Features.EnablePreflight && isNewResource && preflight.IsSupported(plan.Type.ValueString(), plan.ParentID.ValueString()) {
+	if r.ProviderData.Features.EnablePreflight && isNewResource {
 		parentId := plan.ParentID.ValueString()
 		if parentId == "" {
 			placeholder, err := preflight.ParentIdPlaceholder(resourceDef, r.ProviderData.Account.GetSubscriptionId())
@@ -585,7 +580,7 @@ func (r *AzapiResource) ModifyPlan(ctx context.Context, request resource.ModifyP
 }
 
 func (r *AzapiResource) Create(ctx context.Context, request resource.CreateRequest, response *resource.CreateResponse) {
-	r.CreateUpdate(ctx, request.Plan, &response.State, &response.Diagnostics)
+	r.CreateUpdate(ctx, request.Config, request.Plan, &response.State, &response.Diagnostics, response.Private)
 }
 
 func (r *AzapiResource) Update(ctx context.Context, request resource.UpdateRequest, response *resource.UpdateResponse) {
@@ -603,11 +598,12 @@ func (r *AzapiResource) Update(ctx context.Context, request resource.UpdateReque
 		return
 	}
 	tflog.Debug(ctx, "azapi_resource.CreateUpdate proceeding with external request as no skippable changes were detected")
-	r.CreateUpdate(ctx, request.Plan, &response.State, &response.Diagnostics)
+	r.CreateUpdate(ctx, request.Config, request.Plan, &response.State, &response.Diagnostics, response.Private)
 }
 
-func (r *AzapiResource) CreateUpdate(ctx context.Context, requestPlan tfsdk.Plan, responseState *tfsdk.State, diagnostics *diag.Diagnostics) {
-	var plan, state *AzapiResourceModel
+func (r *AzapiResource) CreateUpdate(ctx context.Context, requestConfig tfsdk.Config, requestPlan tfsdk.Plan, responseState *tfsdk.State, diagnostics *diag.Diagnostics, privateData PrivateData) {
+	var config, plan, state *AzapiResourceModel
+	diagnostics.Append(requestConfig.Get(ctx, &config)...)
 	diagnostics.Append(requestPlan.Get(ctx, &plan)...)
 	diagnostics.Append(responseState.Get(ctx, &state)...)
 	if diagnostics.HasError() {
@@ -637,22 +633,12 @@ func (r *AzapiResource) CreateUpdate(ctx context.Context, requestPlan tfsdk.Plan
 			return
 		}
 	}
-	var client clients.Requester
-	client = r.ProviderData.ResourceClient
-	if !plan.Retry.IsNull() {
-		regexps := clients.StringSliceToRegexpSliceMust(plan.Retry.GetErrorMessages())
-		bkof := backoff.NewExponentialBackOff(
-			backoff.WithInitialInterval(plan.Retry.GetIntervalSecondsAsDuration()),
-			backoff.WithMaxInterval(plan.Retry.GetMaxIntervalSecondsAsDuration()),
-			backoff.WithMultiplier(plan.Retry.GetMultiplier()),
-			backoff.WithRandomizationFactor(plan.Retry.GetRandomizationFactor()),
-			backoff.WithMaxElapsedTime(timeout),
-		)
-		tflog.Debug(ctx, "azapi_resource.CreateUpdate is using retry")
-		client = r.ProviderData.ResourceClient.WithRetry(bkof, regexps, nil, nil)
-	}
+
 	ctx, cancel := context.WithTimeout(ctx, timeout)
 	defer cancel()
+
+	// Ensure the context deadline has been set before calling ConfigureClientWithCustomRetry().
+	client := r.ProviderData.ResourceClient.ConfigureClientWithCustomRetry(ctx, plan.Retry, false)
 
 	if isNewResource {
 		// check if the resource already exists using the non-retry client to avoid issue where user specifies
@@ -679,6 +665,12 @@ func (r *AzapiResource) CreateUpdate(ctx context.Context, requestPlan tfsdk.Plan
 	if diagnostics.Append(expandBody(body, *plan)...); diagnostics.HasError() {
 		return
 	}
+	SensitiveBody := make(map[string]interface{})
+	if err := unmarshalBody(config.SensitiveBody, &SensitiveBody); err != nil {
+		diagnostics.AddError("Invalid sensitive_body", fmt.Sprintf(`The argument "sensitive_body" is invalid: %s`, err.Error()))
+		return
+	}
+	body = utils.MergeObject(body, SensitiveBody).(map[string]interface{})
 
 	if !isNewResource {
 		// handle the case that identity block was once set, now it's removed
@@ -742,21 +734,9 @@ func (r *AzapiResource) CreateUpdate(ctx context.Context, requestPlan tfsdk.Plan
 		diagnostics.AddError("Failed to create/update resource", fmt.Errorf("creating/updating %s: %+v", id, err).Error())
 		return
 	}
-	// Create a new retry client to handle specific case of transient 404 or empty body after resource creation
-	clientGetAfterPut := r.ProviderData.ResourceClient.WithRetry(
-		backoff.NewExponentialBackOff(
-			backoff.WithInitialInterval(5*time.Second),
-			backoff.WithMaxInterval(30*time.Second),
-			backoff.WithMaxElapsedTime(RetryGetAfterPut()),
-		),
-		plan.Retry.GetErrorMessagesRegex(),
-		[]int{404},
-		[]func(d interface{}) bool{
-			func(d interface{}) bool {
-				return d == nil
-			},
-		},
-	)
+
+	clientGetAfterPut := r.ProviderData.ResourceClient.ConfigureClientWithCustomRetry(ctx, plan.Retry, true)
+
 	tflog.Debug(ctx, "azapi_resource.CreateUpdate get resource after creation")
 	responseBody, err := clientGetAfterPut.Get(ctx, id.AzureResourceId, id.ApiVersion, clients.NewRequestOptions(AsMapOfString(plan.ReadHeaders), AsMapOfLists(plan.ReadQueryParameters)))
 	if err != nil {
@@ -798,6 +778,13 @@ func (r *AzapiResource) CreateUpdate(ctx context.Context, requestPlan tfsdk.Plan
 		}
 	}
 	diagnostics.Append(responseState.Set(ctx, plan)...)
+
+	writeOnlyBytes, err := dynamic.ToJSON(config.SensitiveBody)
+	if err != nil {
+		diagnostics.AddError("Invalid sensitive_body", err.Error())
+		return
+	}
+	diagnostics.Append(ephemeralBodyPrivateMgr.Set(ctx, privateData, writeOnlyBytes)...)
 }
 
 func (r *AzapiResource) Read(ctx context.Context, request resource.ReadRequest, response *resource.ReadResponse) {
@@ -812,6 +799,7 @@ func (r *AzapiResource) Read(ctx context.Context, request resource.ReadRequest, 
 		return
 	}
 
+	// Ensure the context deadline has been set before calling ConfigureClientWithCustomRetry().
 	ctx, cancel := context.WithTimeout(ctx, readTimeout)
 	defer cancel()
 
@@ -823,20 +811,7 @@ func (r *AzapiResource) Read(ctx context.Context, request resource.ReadRequest, 
 
 	ctx = tflog.SetField(ctx, "resource_id", id.ID())
 
-	var client clients.Requester
-	client = r.ProviderData.ResourceClient
-	if !model.Retry.IsNull() && !model.Retry.IsUnknown() {
-		regexps := clients.StringSliceToRegexpSliceMust(model.Retry.GetErrorMessages())
-		bkof := backoff.NewExponentialBackOff(
-			backoff.WithInitialInterval(model.Retry.GetIntervalSecondsAsDuration()),
-			backoff.WithMaxInterval(model.Retry.GetMaxIntervalSecondsAsDuration()),
-			backoff.WithMultiplier(model.Retry.GetMultiplier()),
-			backoff.WithRandomizationFactor(model.Retry.GetRandomizationFactor()),
-			backoff.WithMaxElapsedTime(readTimeout),
-		)
-		tflog.Debug(ctx, "azapi_resource.Read is using retry")
-		client = r.ProviderData.ResourceClient.WithRetry(bkof, regexps, nil, nil)
-	}
+	client := r.ProviderData.ResourceClient.ConfigureClientWithCustomRetry(ctx, model.Retry, false)
 
 	responseBody, err := client.Get(ctx, id.AzureResourceId, id.ApiVersion, clients.NewRequestOptions(AsMapOfString(model.ReadHeaders), AsMapOfLists(model.ReadQueryParameters)))
 	if err != nil {
@@ -946,6 +921,7 @@ func (r *AzapiResource) Read(ctx context.Context, request resource.ReadRequest, 
 			return
 		}
 		state.Body = payload
+		response.Diagnostics.Append(response.Private.SetKey(ctx, FlagMoveState, []byte("false"))...)
 	}
 
 	response.Diagnostics.Append(response.State.Set(ctx, state)...)
@@ -971,23 +947,11 @@ func (r *AzapiResource) Delete(ctx context.Context, request resource.DeleteReque
 		return
 	}
 
-	var client clients.Requester
-	client = r.ProviderData.ResourceClient
-	if !model.Retry.IsNull() && !model.Retry.IsUnknown() {
-		regexps := clients.StringSliceToRegexpSliceMust(model.Retry.GetErrorMessages())
-		bkof := backoff.NewExponentialBackOff(
-			backoff.WithInitialInterval(model.Retry.GetIntervalSecondsAsDuration()),
-			backoff.WithMaxInterval(model.Retry.GetMaxIntervalSecondsAsDuration()),
-			backoff.WithMultiplier(model.Retry.GetMultiplier()),
-			backoff.WithRandomizationFactor(model.Retry.GetRandomizationFactor()),
-			backoff.WithMaxElapsedTime(deleteTimeout),
-		)
-		tflog.Debug(ctx, "azapi_resource.Delete is using retry")
-		client = r.ProviderData.ResourceClient.WithRetry(bkof, regexps, nil, nil)
-	}
-
+	// Ensure the context deadline has been set before calling ConfigureClientWithCustomRetry().
 	ctx, cancel := context.WithTimeout(ctx, deleteTimeout)
 	defer cancel()
+
+	client := r.ProviderData.ResourceClient.ConfigureClientWithCustomRetry(ctx, model.Retry, false)
 
 	lockIds := AsStringList(model.Locks)
 	slices.Sort(lockIds)
@@ -1094,9 +1058,16 @@ func (r *AzapiResource) MoveState(ctx context.Context) []resource.StateMover {
 					response.Diagnostics.AddError("Invalid source state", "The source state does not contain an id")
 					return
 				}
-				id, err := parse.ResourceID(requestID)
+
+				azureId, err := parse.AzurermIdToAzureId(request.SourceTypeName, requestID)
 				if err != nil {
 					response.Diagnostics.AddError("Invalid Resource ID", fmt.Errorf("parsing Resource ID %q: %+v", requestID, err).Error())
+					return
+				}
+
+				id, err := parse.ResourceID(azureId)
+				if err != nil {
+					response.Diagnostics.AddError("Invalid Resource ID", fmt.Errorf("parsing Resource ID %q: %+v", azureId, err).Error())
 					return
 				}
 
@@ -1125,55 +1096,104 @@ func (r *AzapiResource) nameWithDefaultNaming(config types.String) (types.String
 	}
 }
 
-func (r *AzapiResource) tagsWithDefaultTags(config types.Map, body map[string]interface{}, state *AzapiResourceModel, resourceDef *aztypes.ResourceType) types.Map {
-	if config.IsNull() {
-		switch {
-		case body["tags"] != nil:
-			return tags.FlattenTags(body["tags"])
-		case len(r.ProviderData.Features.DefaultTags) != 0 && canResourceHaveProperty(resourceDef, "tags"):
-			defaultTags := r.ProviderData.Features.DefaultTags
-			if state == nil || state.Tags.IsNull() {
-				return tags.FlattenTags(defaultTags)
-			} else {
-				currentTags := tags.ExpandTags(state.Tags)
-				if !reflect.DeepEqual(currentTags, defaultTags) {
-					return tags.FlattenTags(defaultTags)
-				} else {
-					return state.Tags
-				}
+func (r *AzapiResource) tagsWithDefaultTags(config types.Map, state *AzapiResourceModel, body types.Dynamic, resourceDef *aztypes.ResourceType) types.Map {
+	// 1. use the tags in config if it's not null
+	if !config.IsNull() {
+		return config
+	}
+
+	// 2. use the tags in body if it's not null
+	if !body.IsNull() && !body.IsUnknown() && !body.IsUnderlyingValueNull() && !body.IsUnderlyingValueUnknown() {
+		if bodyObject, ok := body.UnderlyingValue().(types.Object); ok {
+			if v, ok := bodyObject.Attributes()["tags"]; ok && v != nil {
+				return tags.FlattenTags(v)
 			}
-		// To suppress the diff of config: tags = null and state: tags = {}
-		case state != nil && !state.Tags.IsUnknown() && len(state.Tags.Elements()) == 0:
-			return state.Tags
 		}
 	}
-	return config
+
+	// 3. use the default tags if it's not null and the resource supports tags
+	if len(r.ProviderData.Features.DefaultTags) != 0 && canResourceHaveProperty(resourceDef, "tags") {
+		defaultTags := r.ProviderData.Features.DefaultTags
+
+		// if it's a new resource or the tags in state is null, use the default tags
+		if state == nil || state.Tags.IsNull() {
+			return tags.FlattenTags(defaultTags)
+		}
+
+		// if the tags in state is not null and the tags in state is not equal to the default tags, use the default tags
+		currentTags := tags.ExpandTags(state.Tags)
+		if !reflect.DeepEqual(currentTags, defaultTags) {
+			return tags.FlattenTags(defaultTags)
+		}
+
+		return state.Tags
+	}
+
+	// 4. To suppress the diff of config: tags = null and state: tags = {}
+	if state != nil && !state.Tags.IsUnknown() && len(state.Tags.Elements()) == 0 {
+		return state.Tags
+	}
+
+	// 5. return null if all the above cases are null
+	return types.MapNull(types.StringType)
 }
 
-func (r *AzapiResource) locationWithDefaultLocation(config types.String, body map[string]interface{}, state *AzapiResourceModel, resourceDef *aztypes.ResourceType) types.String {
-	if config.IsNull() {
-		switch {
-		case body["location"] != nil:
-			return types.StringValue(body["location"].(string))
-		case len(r.ProviderData.Features.DefaultLocation) != 0 && canResourceHaveProperty(resourceDef, "location"):
-			defaultLocation := r.ProviderData.Features.DefaultLocation
-			if state == nil || state.Location.IsNull() {
-				return types.StringValue(defaultLocation)
-			} else {
-				currentLocation := state.Location.ValueString()
-				if location.Normalize(currentLocation) != location.Normalize(defaultLocation) {
-					return types.StringValue(defaultLocation)
-				} else {
-					return state.Location
+func (r *AzapiResource) locationWithDefaultLocation(configLocation types.String, planLocation types.String, state *AzapiResourceModel, body types.Dynamic, resourceDef *aztypes.ResourceType) types.String {
+	// location field has a field level plan modifier which suppresses the diff if the location is not actually changed
+	config := planLocation
+	// For the following cases, we need to use the location in config as the specified location
+	// case 1. To create a new resource, the location is not specified in config, then the planned location will be unknown
+	// case 2. To update a resource, the location is not specified in config, then the planned location will be the state location
+	if config.IsUnknown() || configLocation.IsNull() {
+		config = configLocation
+	}
+
+	// 1. use the location in config if it's not null
+	if !config.IsNull() {
+		return config
+	}
+
+	// 2. use the location in body if it's not null
+	if !body.IsNull() && !body.IsUnknown() && !body.IsUnderlyingValueNull() && !body.IsUnderlyingValueUnknown() {
+		if bodyObject, ok := body.UnderlyingValue().(types.Object); ok {
+			if v, ok := bodyObject.Attributes()["location"]; ok && v != nil {
+				if strV, ok := v.(types.String); ok {
+					return strV
 				}
 			}
-		// To suppress the diff of config: location = null and state: location = ""
-		// This case happens when upgrading resources which doesn't support location from terraform-plugin-sdk built azapi provider
-		case state != nil && !state.Location.IsUnknown() && state.Location.ValueString() == "":
-			return state.Location
 		}
 	}
-	return config
+
+	// 3. use the state location if it's not specified in config but returned by the API
+	if state != nil && state.Location.ValueString() != "" {
+		return state.Location
+	}
+
+	// 4. use the default location if it's not null and the resource supports location
+	if len(r.ProviderData.Features.DefaultLocation) != 0 && canResourceHaveProperty(resourceDef, "location") {
+		defaultLocation := r.ProviderData.Features.DefaultLocation
+
+		// if it's a new resource or the location in state is null, use the default location
+		if state == nil || state.Location.IsNull() {
+			return types.StringValue(defaultLocation)
+		}
+
+		// if the location in state is not null and the location in state is not equal to the default location, use the default location
+		currentLocation := state.Location.ValueString()
+		if location.Normalize(currentLocation) != location.Normalize(defaultLocation) {
+			return types.StringValue(defaultLocation)
+		}
+
+		return state.Location
+	}
+
+	// 5. To suppress the diff of config: location = null and state: location = ""
+	if state != nil && !state.Location.IsUnknown() && state.Location.ValueString() == "" {
+		return state.Location
+	}
+
+	// 6. return null if all the above cases are null
+	return types.StringNull()
 }
 
 func (r *AzapiResource) defaultAzapiResourceModel() AzapiResourceModel {
@@ -1237,16 +1257,22 @@ func expandBody(body map[string]interface{}, model AzapiResourceModel) diag.Diag
 	return diag.Diagnostics{}
 }
 
-func validateDuplicatedDefinitions(model *AzapiResourceModel, body map[string]interface{}) diag.Diagnostics {
+func validateDuplicatedDefinitions(model *AzapiResourceModel, body types.Dynamic, attributePath string) diag.Diagnostics {
 	diags := diag.Diagnostics{}
-	if !model.Tags.IsNull() && !model.Tags.IsUnknown() && body["tags"] != nil {
-		diags.AddError("Invalid configuration", `can't specify both the argument "tags" and "tags" in the argument "body"`)
+	if body.IsNull() || body.IsUnknown() || body.IsUnderlyingValueNull() || body.IsUnderlyingValueUnknown() {
+		return diags
 	}
-	if !model.Location.IsNull() && !model.Location.IsUnknown() && body["location"] != nil {
-		diags.AddError("Invalid configuration", `can't specify both the argument "location" and "location" in the argument "body"`)
-	}
-	if !model.Identity.IsNull() && !model.Identity.IsUnknown() && body["identity"] != nil {
-		diags.AddError("Invalid configuration", `can't specify both the argument "identity" and "identity" in the argument "body"`)
+
+	if bodyObject, ok := body.UnderlyingValue().(types.Object); ok {
+		if !model.Tags.IsNull() && !model.Tags.IsUnknown() && bodyObject.Attributes()["tags"] != nil {
+			diags.AddError("Invalid configuration", fmt.Sprintf(`can't specify both the argument "tags" and "tags" in the argument "%s"`, attributePath))
+		}
+		if !model.Location.IsNull() && !model.Location.IsUnknown() && bodyObject.Attributes()["location"] != nil {
+			diags.AddError("Invalid configuration", fmt.Sprintf(`can't specify both the argument "location" and "location" in the argument "%s"`, attributePath))
+		}
+		if !model.Identity.IsNull() && !model.Identity.IsUnknown() && bodyObject.Attributes()["identity"] != nil {
+			diags.AddError("Invalid configuration", fmt.Sprintf(`can't specify both the argument "identity" and "identity" in the argument "%s"`, attributePath))
+		}
 	}
 	return diags
 }